delete user_transaction file and change main to St
