--- lib/widgets/user_transactions.dart
@@ -1,36 +1,0 @@
-import 'package:flutter/material.dart';
-import './transaction_list.dart';
-import './new_transaction.dart';
-import '../models/transaction.dart';
-
-class UserTransaction extends StatefulWidget {
-  @override
-  _UserTransactionState createState() => _UserTransactionState();
-}
-
-class _UserTransactionState extends State<UserTransaction> {
-  final List<Transaction> _userTransactions = [
-    Transaction(id: 't1', title: 'book', amount: 25.99, date: DateTime.now()),
-    Transaction(id: 't2', title: 'food', amount: 75.99, date: DateTime.now()),
-  ];
-  void _addNewTransaction(String txTitle, double txAmount) {
-    final newTx = Transaction(
-      id: DateTime.now().toString(),
-      title: txTitle,
-      amount: txAmount,
-      date: DateTime.now(),
-    );
-    setState((){
-_userTransactions.add(newTx);
-  });
-  }
-  
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(children: [
-      NewTransaction(_addNewTransaction),
-      TransactionList(_userTransactions),
-    ]);
-  }
-}
